disable the map button while there is no record in the database, modify the output of the title text in the record view from string value to double value, since the 8.62 will be stored as 8.6199999 which is a bug

+++ EasyLife/SharedResultViewController.m
@@ -22,7 +22,7 @@
 @property (strong, nonatomic) CLLocationManager *locationManager;
 @property (strong,nonatomic) CLLocation *location;
 @property (nonatomic) NSInteger locationErrorCode;
-@property float numberOfPeopleToShare;
+@property double numberOfPeopleToShare;
 @end
 
 @implementation SharedResultViewController
